bug/pkg/health: Fix Nil Address Issue in Node Update Mechanism
It is possible for a node with a "<nil>" primary address to
be "added" to the health pkg cache. When this happens subsequent
updates **will not** flush out the bad value and thus it will
persist as a valid status for a node until the entire cilium
daemon is reset.

This fixes the bug by not caching nil values at all.

Signed-off-by: Nate Sweet <nathanjsweet@pm.me>